refactor TeammateDisplay
